@@ -12,13 +12,15 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Markdown.MAML.Parser;
+using Markdown.MAML.Transformer;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Management.Automation;
+using System.Reflection;
 using System.Text.RegularExpressions;
-using Tools.Common.Helpers;
 using Tools.Common.Issues;
 using Tools.Common.Loaders;
 using Tools.Common.Loggers;
@@ -33,6 +35,8 @@ namespace StaticAnalysis.HelpAnalyzer
     {
         private const int MissingHelp = 6050;
         private const int MissingHelpFile = 6000;
+        private const int MissingCmdlet = 7000;
+        private const int PlatyPSSchemaViolation = 8000;
         public HelpAnalyzer()
         {
             Name = "Help Analyzer";
@@ -40,7 +44,7 @@ namespace StaticAnalysis.HelpAnalyzer
         public AnalysisLogger Logger { get; set; }
         public string Name { get; private set; }
 
-// TODO: Remove IfDef code
+        // TODO: Remove IfDef code
 #if !NETSTANDARD
         private AppDomain _appDomain;
 #endif
@@ -92,6 +96,7 @@ namespace StaticAnalysis.HelpAnalyzer
             var helpLogger = Logger.CreateLogger<HelpIssue>("HelpIssues.csv");
             foreach (var baseDirectory in scopes.Where(s => Directory.Exists(Path.GetFullPath(s))))
             {
+                SharedAssemblyLoader.Load(baseDirectory);
                 foreach (var directory in Directory.EnumerateDirectories(Path.GetFullPath(baseDirectory)))
                 {
                     if (modulesToAnalyze != null &&
@@ -104,6 +109,7 @@ namespace StaticAnalysis.HelpAnalyzer
                     var dirs = Directory.EnumerateDirectories(directory);
                     if (dirs != null && dirs.Any(d => string.Equals(Path.GetFileName(d), "help", StringComparison.OrdinalIgnoreCase)))
                     {
+                        Console.WriteLine($"Analyzing module under {directory} ...");
                         AnalyzeMarkdownHelp(scopes, directory, helpLogger, processedHelpFiles, savedDirectory);
                     }
                 }
@@ -154,7 +160,7 @@ namespace StaticAnalysis.HelpAnalyzer
                     h.Assembly = cmdletFileName;
                 }, "Cmdlet");
 
-// TODO: Remove IfDef
+                // TODO: Remove IfDef
 #if NETSTANDARD
                 var proxy = new CmdletLoader();
 #else
@@ -165,7 +171,7 @@ namespace StaticAnalysis.HelpAnalyzer
                 var helpRecords = CmdletHelpParser.GetHelpTopics(helpFile, helpLogger);
                 ValidateHelpRecords(cmdlets, helpRecords, helpLogger);
                 helpLogger.Decorator.Remove("Cmdlet");
-// TODO: Remove IfDef code
+                // TODO: Remove IfDef code
 #if !NETSTANDARD
                 AppDomain.Unload(_appDomain);
 #endif
@@ -201,6 +207,8 @@ namespace StaticAnalysis.HelpAnalyzer
             }
 
             var helpFiles = Directory.EnumerateFiles(helpFolder, "*.md").Select(Path.GetFileNameWithoutExtension).ToList();
+            // Assume all cmdlets markdown file following format of VERB-AzResource. Dash is required.
+            helpFiles = helpFiles.Where(c => c.Contains("-")).ToList();
             if (!helpFiles.Any()) return;
 
             Directory.SetCurrentDirectory(directory);
@@ -218,77 +226,96 @@ namespace StaticAnalysis.HelpAnalyzer
             var psd1 = manifestFiles.FirstOrDefault();
             var parentDirectory = Directory.GetParent(psd1).FullName;
             var psd1FileName = Path.GetFileName(psd1);
-            var powershell = PowerShell.Create();
-            powershell.AddScript("Import-LocalizedData -BaseDirectory " + parentDirectory +
-                                 " -FileName " + psd1FileName +
-                                 " -BindingVariable ModuleMetadata; $ModuleMetadata.NestedModules; $ModuleMetadata.RequiredModules | % { $_[\"ModuleName\"] };");
-            var cmdletResult = powershell.Invoke();
-            var nestedModules = cmdletResult.Where(c => c.ToString().StartsWith(".")).Select(c => c.ToString().Substring(2));
-            var requiredModules = cmdletResult.Where(c => !c.ToString().StartsWith(".")).Select(c => c.ToString()).ToList();
-            if (nestedModules.Any())
+
+            var allCmdlets = new List<CmdletMetadata>();
+
+
+            string moduleName = psd1FileName.Replace(".psd1", "");
+
+            Console.WriteLine(directory);
+            Directory.SetCurrentDirectory(directory);
+
+            processedHelpFiles.Add(moduleName);
+
+            helpLogger.Decorator.AddDecorator(h =>
             {
-                Directory.SetCurrentDirectory(directory);
+                h.HelpFile = moduleName;
+                h.Assembly = moduleName;
+            }, "Cmdlet");
+            var module = MetadataLoader.GetModuleMetadata(moduleName);
+            var cmdlets = module.Cmdlets;
+            allCmdlets.AddRange(cmdlets);
+            helpLogger.Decorator.Remove("Cmdlet");
 
-                requiredModules = requiredModules.Join(scopes,
-                        module => 1,
-                        dir => 1,
-                        (module, dir) => Path.Combine(dir, module))
-                    .Where(Directory.Exists)
-                    .ToList();
-
-                requiredModules.Add(directory);
-                var allCmdlets = new List<CmdletMetadata>();
-                foreach (var nestedModule in nestedModules)
-                {
-                    var assemblyFile = Directory.GetFiles(parentDirectory, nestedModule, SearchOption.AllDirectories).FirstOrDefault();
-                    if (!File.Exists(assemblyFile)) continue;
-
-                    var assemblyFileName = Path.GetFileName(assemblyFile);
-                    helpLogger.Decorator.AddDecorator(h =>
-                    {
-                        h.HelpFile = assemblyFileName;
-                        h.Assembly = assemblyFileName;
-                    }, "Cmdlet");
-                    processedHelpFiles.Add(assemblyFileName);
-// TODO: Remove IfDef
-#if NETSTANDARD
-                    var proxy = new CmdletLoader();
-#else
-                    var proxy = EnvironmentHelpers.CreateProxy<CmdletLoader>(directory, out _appDomain);
-#endif
-                    var module = proxy.GetModuleMetadata(assemblyFile, requiredModules);
-                    var cmdlets = module.Cmdlets;
-                    allCmdlets.AddRange(cmdlets);
-                    helpLogger.Decorator.Remove("Cmdlet");
-// TODO: Remove IfDef code
-#if !NETSTANDARD
-                    AppDomain.Unload(_appDomain);
-#endif
-                }
-
-                ValidateHelpRecords(allCmdlets, helpFiles, helpLogger);
-            }
+            ValidateHelpRecords(allCmdlets, helpFiles, helpLogger);
+            ValidateHelpMarkdown(helpFolder, helpFiles, helpLogger);
 
             Directory.SetCurrentDirectory(savedDirectory);
-            
+
         }
 
         private void ValidateHelpRecords(IList<CmdletMetadata> cmdlets, IList<string> helpRecords,
             ReportLogger<HelpIssue> helpLogger)
         {
+            var cmdletDict = new Dictionary<string, CmdletMetadata>();
             foreach (var cmdlet in cmdlets)
             {
+                cmdletDict.Add(cmdlet.Name, cmdlet);
                 if (!helpRecords.Contains(cmdlet.Name, StringComparer.OrdinalIgnoreCase))
                 {
-                    helpLogger.LogRecord(new HelpIssue
+                    HelpIssue issue = new HelpIssue
                     {
                         Target = cmdlet.ClassName,
                         Severity = 1,
                         ProblemId = MissingHelp,
-                        Description = string.Format("Help missing for cmdlet {0} implemented by class {1}",
-                        cmdlet.Name, cmdlet.ClassName),
                         Remediation = string.Format("Add Help record for cmdlet {0} to help file.", cmdlet.Name)
-                    });
+                    };
+                    if (cmdlet.ClassName != null)
+                    {
+                        issue.Description = $"Help missing for cmdlet {cmdlet.Name} implemented by class {cmdlet.ClassName}";
+                    }
+                    else
+                    {
+                        issue.Description = $"Help missing for cmdlet {cmdlet.Name} implemented by functions";
+                    }
+                    helpLogger.LogRecord(issue);
+                }
+            }
+
+            foreach (var helpRecord in helpRecords)
+            {
+                if (!cmdletDict.ContainsKey(helpRecord))
+                {
+                    Console.Error.WriteLine($"Help record {helpRecord} has no cmdlet.");
+                }
+            }
+        }
+
+        private void ValidateHelpMarkdown(string helpFolder, IList<string> helpRecords, ReportLogger<HelpIssue> helpLogger)
+        {
+            foreach (var helpMarkdown in helpRecords)
+            {
+                var file = Path.Combine(helpFolder, helpMarkdown + ".md");
+                var content = File.ReadAllText(file);
+                try
+                {
+                    var parser = new MarkdownParser();
+                    var transformer = new ModelTransformerVersion2();
+                    var markdownModel = parser.ParseString(new[] { content });
+                    var model = transformer.NodeModelToMamlModel(markdownModel).FirstOrDefault();
+                }
+                catch (Exception ex)
+                {
+                    HelpIssue issue = new HelpIssue
+                    {
+                        Target = helpMarkdown,
+                        Severity = 1,
+                        ProblemId = PlatyPSSchemaViolation,
+                        Description = "Help content doesn't conform to PlatyPS Schema definition",
+                        Remediation = string.Format("No.")
+                    };
+                    helpLogger.LogRecord(issue);
+                    Console.Error.WriteLine($"Failed to parse {file} by PlatyPS, {ex.Message}");
                 }
             }
         }
