@@ -65,7 +65,7 @@ namespace StaticAnalysis.HelpAnalyzer
 
         public IReportRecord Parse(string line)
         {
-            var matcher = "\"([^\"]+)\",\"([^\"]+)\",\"([^\"]+)\",\"([^\"]+)\",\"([^\"]+)\",\"([^\"]+)\",\"([^\"]+)\"";
+            var matcher = "\"([^\"]*)\",\"([^\"]*)\",\"([^\"]*)\",\"([^\"]+)\",\"([^\"]+)\",\"([^\"]*)\",\"([^\"]*)\"";
             var match = Regex.Match(line, matcher);
             if (!match.Success || match.Groups.Count < 8)
             {
